Migrate to TS and add tslint/prettier integration, shim with any until I have time for better types...
