+++ src/types/global.d.ts
@@ -1,0 +1,1 @@
+declare const __PATH_PREFIX__: string
