+++ src/types/node_modules.d.ts
@@ -1,0 +1,1 @@
+declare module "typography-theme-wordpress-2016"
