+++ tslint.json
@@ -1,0 +1,12 @@
+{
+  "extends": ["tslint:latest", "tslint-react", "tslint-eslint-rules", "tslint-config-prettier"],
+  "rules": {
+    "prettier": true,
+    "jsx-no-multiline-js": false,
+    "jsx-no-lambda": false,
+    "import-name": false,
+    "interface-name": false,
+    "no-boolean-literal-compare": false,
+    "object-literal-sort-keys": false
+  }
+}
